Moved Assignment and Import to models folder
