+++ examples/model/assignment.py
